创建表后，插入示例数据；增加XxlJobService测试类

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/service/XxlJobService.java
@@ -52,7 +52,7 @@ public interface XxlJobService {
 	 * @param id
 	 * @return
 	 */
-	public ReturnT<String> remove(int id);
+	public ReturnT<String> remove(long id);
 
 	/**
 	 * start job
@@ -60,7 +60,7 @@ public interface XxlJobService {
 	 * @param id
 	 * @return
 	 */
-	public ReturnT<String> start(int id);
+	public ReturnT<String> start(long id);
 
 	/**
 	 * stop job
@@ -68,7 +68,7 @@ public interface XxlJobService {
 	 * @param id
 	 * @return
 	 */
-	public ReturnT<String> stop(int id);
+	public ReturnT<String> stop(long id);
 
 	/**
 	 * dashboard info
